Consolidate primary-key checking into Resource
Move composite primary-key uniqueness checking (previously duplicated
between Resource.enforce_schema's inline pandas logic and
asset_checks.py's standalone Polars implementation) into a single
Resource.check_primary_key() dispatcher, so callers no longer need to
pick between backend-specific methods.

The pandas path now reports duplicate-key and null-key violations
together in one ValueError instead of stopping at whichever is found
first, which was awkward for diagnosing multiple problems at once.
The Polars path, its chunked-checking helpers, and the large-table
chunking table move from asset_checks.py to Resource as private/static
methods, alongside the other schema-derived helpers they depend on.
